requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic-settings>=2.1.0
httpx>=0.26.0
pytest>=8.0.0
black>=24.1.0
ruff>=0.1.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
